Move env setup subcommand into env.go
For now use one file per command and include subcommands in them.
